// src/generated/graphql.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
};

/**
 * Account: A higher order entity created for any addresses which interact with Superfluid contracts.
 *
 */
export type Account = {
  __typename?: 'Account';
  accountTokenSnapshots: Array<AccountTokenSnapshot>;
  createdAtBlockNumber: Scalars['BigInt'];
  createdAtTimestamp: Scalars['BigInt'];
  id: Scalars['ID'];
  inflows: Array<Stream>;
  /**
   * Indicates whether the address/account is a super app.
   *
   */
  isSuperApp: Scalars['Boolean'];
  outflows: Array<Stream>;
  publishedIndexes: Array<Index>;
  receivedTransferEvents: Array<TransferEvent>;
  sentTransferEvents: Array<TransferEvent>;
  subscriptions: Array<IndexSubscription>;
  tokenDowngradedEvents: Array<TokenDowngradedEvent>;
  tokenUpgradedEvents: Array<TokenUpgradedEvent>;
  updatedAtBlockNumber: Scalars['BigInt'];
  updatedAtTimestamp: Scalars['BigInt'];
};
// ...
export type FlowUpdatedEvent = Event & {
  __typename?: 'FlowUpdatedEvent';
  /**
   * Contains the addresses that were impacted by this event:
   * addresses[0] = `token` (supertoken)
   * addresses[1] = `sender`
   * addresses[2] = `receiver`
   *
   */
  addresses: Array<Scalars['Bytes']>;
  blockNumber: Scalars['BigInt'];
  /**
   * The deposit amount put up for the creation of the flow.
   *
   */
  deposit: Scalars['BigInt'];
  /**
   * The address that is executing the flow update transaction.
   * This will be the zero address until the flowOperator feature is live.
   *
   */
  flowOperator: Scalars['Bytes'];
  /**
   * The flow rate per second.
   *
   */
  flowRate: Scalars['BigInt'];
  gasPrice: Scalars['BigInt'];
  gasUsed: Scalars['BigInt'];
  id: Scalars['ID'];
  logIndex: Scalars['BigInt'];
  name: Scalars['String'];
  /**
   * The previous flow rate, the absolute (positive) value.
   *
   */
  oldFlowRate: Scalars['BigInt'];
  order: Scalars['BigInt'];
  /**
   * The address of the flow receiver.
   *
   */
  receiver: Scalars['Bytes'];
  /**
   * The address of the flow sender.
   *
   */
  sender: Scalars['Bytes'];
  /**
   * The stream entity which is being modified.
   *
   */
  stream: Stream;
  timestamp: Scalars['BigInt'];
  /**
   * The address of the `token` (supertoken) being streamed.
   *
   */
  token: Scalars['Bytes'];
  /**
   * The total amount streamed until the timestamp
   * for the Stream entity linked to this event.
   *
   */
  totalAmountStreamedUntilTimestamp: Scalars['BigInt'];
  /**
   * The total (global/account level) flow rate of `receiver` for `token` as of this event.
   *
   */
  totalReceiverFlowRate: Scalars['BigInt'];
  /**
   * The total (global/account level) flow rate of `sender` for `token` as of this event.
   *
   */
  totalSenderFlowRate: Scalars['BigInt'];
  transactionHash: Scalars['Bytes'];
  /**
   * The "type" of the `FlowUpdated` event.
   * 0 = create
   * 1 = update
   * 2 = terminate
   *
   */
  type: Scalars['Int'];
  /**
   * Arbitrary bytes (additional data) passed upon flow creation.
   *
   */
  userData: Scalars['Bytes'];
};
// ...
export type Stream = {
  __typename?: 'Stream';
  createdAtBlockNumber: Scalars['BigInt'];
  createdAtTimestamp: Scalars['BigInt'];
  currentFlowRate: Scalars['BigInt'];
  deposit: Scalars['BigInt'];
  flowUpdatedEvents: Array<FlowUpdatedEvent>;
  /**
   * ID composed of: senderAddress-receiverAddress-tokenAddress-revisionIndex
   *
   */
  id: Scalars['ID'];
  receiver: Account;
  sender: Account;
  streamPeriods: Array<StreamPeriod>;
  /**
   * The amount streamed until `updatedAtTimestamp`/`updatedAtBlock`.
   * The formula to get the current streamed amount is:
   * `streamedUntilUpdatedAt + ((currentTime in seconds) - updatedAtTimestamp) * currentFlowRate`.
   *
   */
  streamedUntilUpdatedAt: Scalars['BigInt'];
  token: Token;
  updatedAtBlockNumber: Scalars['BigInt'];
  updatedAtTimestamp: Scalars['BigInt'];
};
// ...
export type StreamPeriod = {
  __typename?: 'StreamPeriod';
  deposit: Scalars['BigInt'];
  flowRate: Scalars['BigInt'];
  /**
   * ID composed of: streamId - periodRevisionIndex
   *
   */
  id: Scalars['ID'];
  receiver: Account;
  sender: Account;
  startedAtBlockNumber: Scalars['BigInt'];
  startedAtEvent: FlowUpdatedEvent;
  startedAtTimestamp: Scalars['BigInt'];
  stoppedAtBlockNumber?: Maybe<Scalars['BigInt']>;
  stoppedAtEvent?: Maybe<FlowUpdatedEvent>;
  /**
   * Following values are null until the StreamPeriod is terminated
   *
   */
  stoppedAtTimestamp?: Maybe<Scalars['BigInt']>;
  stream: Stream;
  token: Token;
  totalAmountStreamed?: Maybe<Scalars['BigInt']>;
};
// ...
export type Token = {
  __typename?: 'Token';
  createdAtBlockNumber: Scalars['BigInt'];
  createdAtTimestamp: Scalars['BigInt'];
  decimals: Scalars['Int'];
  /**
   * ID: the token address
   *
   */
  id: Scalars['ID'];
  /**
   * A boolean indicating whether the token is a part of our resolver list.
   *
   */
  isListed: Scalars['Boolean'];
  /**
   * A boolean indicating whether the token is a NativeAssetSuperToken.
   *
   */
  isNativeAssetSuperToken: Scalars['Boolean'];
  isSuperToken: Scalars['Boolean'];
  name: Scalars['String'];
  symbol: Scalars['String'];
  /**
   * The address of the underlying ERC20 token (zero address for non-ERC20WrapperSuperToken's)
   *
   */
  underlyingAddress: Scalars['Bytes'];
  /**
   * The underlying ERC20 token for a ERC20WrapperSuperToken otherwise null.
   *
   */
  underlyingToken?: Maybe<Token>;
};
// ...
export type GetActiveStreamsQuery = { __typename?: 'Query', streams: Array<{ __typename?: 'Stream', id: string }> };

export type GetFlowRateQueryVariables = Exact<{
  sender: Scalars['String'];
  receiver: Scalars['String'];
}>;


export type GetFlowRateQuery = { __typename?: 'Query', streams: Array<{ __typename?: 'Stream', id: string, currentFlowRate: any, updatedAtTimestamp: any, streamedUntilUpdatedAt: any }> };


export const GetActiveStreams = gql`
    query getActiveStreams($sender: String!, $receiver: String!) {
  streams(
    orderBy: createdAtTimestamp
    orderDirection: desc
    first: 1
    where: {receiver: $receiver, sender: $sender, currentFlowRate_gt: 0}
  ) {
    id
  }
}
    `;